refactor: move active note procedures from the miden::note into miden::active_note
