+++ crates/miden-lib/asm/note_scripts/SWAP.masm
@@ -1,4 +1,4 @@
-use.miden::note
+use.miden::active_note
 use.miden::output_note
 use.miden::contracts::wallets::basic->wallet
 
@@ -46,7 +46,7 @@ begin
     # --- create a payback note with the requested asset ----------------
 
     # store note inputs into memory starting at address 0
-    push.0 exec.note::get_inputs
+    push.0 exec.active_note::get_inputs
     # => [num_inputs, inputs_ptr]
 
     # make sure the number of inputs is 12
@@ -89,7 +89,7 @@ begin
     # --- move assets from the SWAP note into the account -------------------------
 
     # store the number of note assets to memory starting at address 0
-    push.0 exec.note::get_assets
+    push.0 exec.active_note::get_assets
     # => [num_assets, ptr, pad(12)]
 
     # make sure the number of assets is 1
